Fix: Improve lock script

+++ i3/.config/i3/scripts/lock.sh
@@ -9,8 +9,7 @@ ICON=/tmp/icon.png
 
 scrot $IMAGE
 convert $IMAGE -scale 10% -scale 1000% $IMAGE
-cp $ORIG_ICON $ICON
-convert $ICON -scale 50% $ICON
+[ -f $ICON ] || cp $ORIG_ICON $ICON
 [ -f $TEXT ] || {
     convert -size 3000x60 xc:white -font Fira-Sans -pointsize 26 -fill black -gravity center -annotate +0+0 'Type password to unlock' $TEXT;
     convert $TEXT -alpha set -channel A -evaluate set 50% $TEXT;
